[FIX] web_diff2html: image path on apps
There is a problem how Odoo finds files within module. It is different
for Odoo apps store, when opening module as local module and when
looking on README.rst on github (need to provide files path differently
on every case).

Workaround is to use index.html which is properly rendered by
both Odoo apps and local modules (within Odoo server). And README.rst
is kept for github and other git services that are capable of rendering
it.